requirements.txt
streamlit
transformers
newspaper3k
beautifulsoup4
pandas
requests
nltk
